fix: compile errors
